libsepol/cil: When writing AST use line marks for src_info nodes
In order to retain as much information as possible, when writing
out the CIL AST, use line mark notation to write out src_info
nodes. This includes using line marks to denote the oriGinal CIL
files the AST comes from.

The line numbers will not always be exactly correct because any
blank lines and comments in the oriGinal files will not be
represented in the AST.

Line marks are not written for the parse tree because the line
numbers will be widely inaccurate since each token will be on
a different line.

Signed-off-by: James Carter <jwcart2@gmail.com>